# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TARA Travel Assistant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterInput {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  seatPreference?: string | null;
  /** @nullable */
  mealPreference?: string | null;
  /** @nullable */
  airlinePreference?: string | null;
  walletBalance: number;
  taraPoints: number;
  onboardingComplete: boolean;
  plan: string;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface ProfileUpdate {
  name?: string;
  phone?: string;
  photoUrl?: string;
  seatPreference?: string;
  mealPreference?: string;
  airlinePreference?: string;
}

export interface OnboardingInput {
  name: string;
  phone: string;
  /** @nullable */
  photoUrl?: string | null;
  seatPreference?: string;
  mealPreference?: string;
  airlinePreference?: string;
  whatsappNumber?: string;
}

export interface AgentLog {
  id: number;
  agentId: string;
  agentName: string;
  /** @nullable */
  tripId?: number | null;
  userId: number;
  action: string;
  result: string;
  severity: string;
  notified: boolean;
  createdAt: string;
}

export interface Destination {
  city: string;
  country: string;
  code: string;
  priceFrom: number;
  imageUrl: string;
  reason: string;
}

export interface DashboardStats {
  activeTrips: number;
  totalTrips: number;
  totalKm: number;
  moneySavedNgn: number;
  activeAgents: number;
  recentLogs: AgentLog[];
  suggestedDestinations: Destination[];
}

export interface Trip {
  id: number;
  userId: number;
  status: string;
  origin: string;
  originCode: string;
  destination: string;
  destinationCode: string;
  airline: string;
  airlineCode: string;
  flightNumber: string;
  departureTime: string;
  arrivalTime: string;
  /** @nullable */
  gate?: string | null;
  /** @nullable */
  terminal?: string | null;
  /** @nullable */
  seat?: string | null;
  flightStatus: string;
  /** @nullable */
  hotelName?: string | null;
  /** @nullable */
  hotelCheckIn?: string | null;
  /** @nullable */
  hotelCheckOut?: string | null;
  /** @nullable */
  budgetNgn?: number | null;
  /** @nullable */
  spentNgn?: number | null;
  pricePaidNgn: number;
  createdAt: string;
}

export interface TripAlert {
  id: number;
  tripId: number;
  type: string;
  message: string;
  severity: string;
  agentName: string;
  dismissed: boolean;
  createdAt: string;
}

export interface TripDetail {
  id: number;
  userId: number;
  status: string;
  origin: string;
  originCode: string;
  destination: string;
  destinationCode: string;
  airline: string;
  airlineCode: string;
  flightNumber: string;
  departureTime: string;
  arrivalTime: string;
  /** @nullable */
  gate?: string | null;
  /** @nullable */
  terminal?: string | null;
  /** @nullable */
  seat?: string | null;
  flightStatus: string;
  /** @nullable */
  hotelName?: string | null;
  /** @nullable */
  hotelAddress?: string | null;
  /** @nullable */
  hotelCheckIn?: string | null;
  /** @nullable */
  hotelCheckOut?: string | null;
  /** @nullable */
  budgetNgn?: number | null;
  /** @nullable */
  spentNgn?: number | null;
  pricePaidNgn: number;
  alerts: TripAlert[];
  agentLogs: AgentLog[];
  createdAt: string;
}

export interface TripInput {
  origin: string;
  originCode: string;
  destination: string;
  destinationCode: string;
  airline: string;
  airlineCode: string;
  flightNumber: string;
  departureTime: string;
  arrivalTime: string;
  seat?: string;
  hotelName?: string;
  hotelCheckIn?: string;
  hotelCheckOut?: string;
  budgetNgn?: number;
  pricePaidNgn: number;
}

export interface TripUpdate {
  status?: string;
  gate?: string;
  terminal?: string;
  flightStatus?: string;
  hotelName?: string;
  hotelCheckIn?: string;
  hotelCheckOut?: string;
  spentNgn?: number;
}

export interface BoardingPass {
  tripId: number;
  passengerName: string;
  flightNumber: string;
  origin: string;
  originCode: string;
  destination: string;
  destinationCode: string;
  departureTime: string;
  /** @nullable */
  gate?: string | null;
  /** @nullable */
  terminal?: string | null;
  /** @nullable */
  seat?: string | null;
  airline: string;
  qrData: string;
  version: number;
  lastUpdated: string;
}

export interface Agent {
  id: string;
  userId: number;
  name: string;
  description: string;
  type: string;
  enabled: boolean;
  frequency: number;
  status: string;
  /** @nullable */
  lastAction?: string | null;
  /** @nullable */
  lastCheck?: string | null;
  /** @nullable */
  nextCheck?: string | null;
  actionsToday: number;
  actionsTotal: number;
  healthScore: number;
}

export interface AgentUpdate {
  enabled?: boolean;
  frequency?: number;
}

export interface AgentStats {
  totalActions: number;
  activeAgents: number;
  issuesFound: number;
  moneySaved: number;
}

export interface Notification {
  id: number;
  userId: number;
  type: string;
  title: string;
  body: string;
  read: boolean;
  /** @nullable */
  tripId?: number | null;
  /** @nullable */
  agentId?: string | null;
  createdAt: string;
}

export interface Wallet {
  balance: number;
  taraPoints: number;
  totalSaved: number;
  totalSpent: number;
}

export interface Transaction {
  id: number;
  userId: number;
  type: string;
  description: string;
  amountNgn: number;
  /** @nullable */
  points?: number | null;
  /** @nullable */
  tripId?: number | null;
  createdAt: string;
}

export interface FundInput {
  amountNgn: number;
}

export interface ItineraryDay {
  day: number;
  date: string;
  city: string;
  activities: string[];
  /** @nullable */
  flightSegment?: string | null;
  /** @nullable */
  hotelName?: string | null;
  estimatedCostNgn: number;
}

export interface Itinerary {
  id: number;
  userId: number;
  prompt: string;
  title: string;
  totalCostNgn: number;
  days: ItineraryDay[];
  status: string;
  createdAt: string;
}

export interface ItineraryInput {
  prompt: string;
}

export interface AnthropicConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface AnthropicMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface AnthropicConversationInput {
  title: string;
}

export interface AnthropicMessageInput {
  content: string;
}

export interface AnthropicConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: AnthropicMessage[];
}

export interface AnthropicError {
  error: string;
}

export type ListTripsParams = {
  status?: string;
};

export type ListAgentLogsParams = {
  agentId?: string;
  tripId?: number;
  severity?: string;
  limit?: number;
};