# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TARA Travel Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── USERS ────────────────────────────────────────────────────────────────
...
  /agents/{agentId}:
    get:
      operationId: getAgent
      tags: [agents]
      summary: Get a single agent
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
    patch:
      operationId: updateAgent
      tags: [agents]
      summary: Update agent settings
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
...
  /agent-logs:
    get:
      operationId: listAgentLogs
      tags: [agentLogs]
      summary: List agent activity logs
      parameters:
        - name: agentId
          in: query
          schema:
            type: string
        - name: tripId
          in: query
          schema:
            type: integer
        - name: severity
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgentLog"

  # ─── NOTIFICATIONS ────────────────────────────────────────────────────────
...
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
  /notifications/{id}:
    delete:
      operationId: deleteNotification
      tags: [notifications]
      summary: Delete a notification
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── WALLET ───────────────────────────────────────────────────────────────
...
  /itineraries:
    get:
      operationId: listItineraries
      tags: [itineraries]
      summary: List saved itineraries
      responses:
        "200":
          description: Itineraries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Itinerary"
    post:
      operationId: createItinerary
      tags: [itineraries]
      summary: Generate an itinerary from a prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItineraryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Itinerary"
  /itineraries/{id}:
    get:
      operationId: getItinerary
      tags: [itineraries]
      summary: Get a single itinerary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Itinerary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Itinerary"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteItinerary
      tags: [itineraries]
      summary: Delete an itinerary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── ANTHROPIC ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    RegisterInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
      required:
        - name
        - email
        - password
    LoginInput:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        seatPreference:
          type: ["string", "null"]
        mealPreference:
          type: ["string", "null"]
        airlinePreference:
          type: ["string", "null"]
        walletBalance:
          type: number
        taraPoints:
          type: integer
        onboardingComplete:
          type: boolean
        plan:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - walletBalance
        - taraPoints
        - onboardingComplete
        - plan
        - createdAt
...
    Agent:
      type: object
      properties:
        id:
          type: string
        userId:
          type: integer
        name:
          type: string
        description:
          type: string
        type:
          type: string
        enabled:
          type: boolean
        frequency:
          type: integer
        status:
          type: string
        lastAction:
          type: ["string", "null"]
        lastCheck:
          type: ["string", "null"]
        nextCheck:
          type: ["string", "null"]
        actionsToday:
          type: integer
        actionsTotal:
          type: integer
        healthScore:
          type: integer
      required:
        - id
        - userId
        - name
        - description
        - type
        - enabled
        - frequency
        - status
        - actionsToday
        - actionsTotal
        - healthScore
    AgentUpdate:
      type: object
      properties:
        enabled:
          type: boolean
        frequency:
          type: integer
...
    AgentLog:
      type: object
      properties:
        id:
          type: integer
        agentId:
          type: string
        agentName:
          type: string
        tripId:
          type: ["integer", "null"]
        userId:
          type: integer
        action:
          type: string
        result:
          type: string
        severity:
          type: string
        notified:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - agentId
        - agentName
        - userId
        - action
        - result
        - severity
        - notified
        - createdAt

    # ─── NOTIFICATIONS ─────────────────────────────────────────────────────
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
        title:
          type: string
        body:
          type: string
        read:
          type: boolean
        tripId:
          type: ["integer", "null"]
        agentId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - type
        - title
        - body
        - read
        - createdAt

    # ─── WALLET ────────────────────────────────────────────────────────────
...
    Itinerary:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        prompt:
          type: string
        title:
          type: string
        totalCostNgn:
          type: integer
        days:
          type: array
          items:
            $ref: "#/components/schemas/ItineraryDay"
        status:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - prompt
        - title
        - totalCostNgn
        - days
        - status
        - createdAt
    ItineraryDay:
      type: object
      properties:
        day:
          type: integer
        date:
          type: string
        city:
          type: string
        activities:
          type: array
          items:
            type: string
        flightSegment:
          type: ["string", "null"]
        hotelName:
          type: ["string", "null"]
        estimatedCostNgn:
          type: integer
      required:
        - day
        - date
        - city
        - activities
        - estimatedCostNgn
    ItineraryInput:
      type: object
      properties:
        prompt:
          type: string
      required:
        - prompt

    # ─── ANTHROPIC ─────────────────────────────────────────────────────────
...