# lib/api-zod/src/generated/types/agentUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TARA Travel Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface AgentUpdate {
  enabled?: boolean;
  frequency?: number;
}