# lib/api-zod/src/generated/types/itinerary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TARA Travel Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { ItineraryDay } from "./itineraryDay";

export interface Itinerary {
  id: number;
  userId: number;
  prompt: string;
  title: string;
  totalCostNgn: number;
  days: ItineraryDay[];
  status: string;
  createdAt: Date;
}