# lib/api-zod/src/generated/types/itineraryDay.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TARA Travel Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface ItineraryDay {
  day: number;
  date: string;
  city: string;
  activities: string[];
  /** @nullable */
  flightSegment?: string | null;
  /** @nullable */
  hotelName?: string | null;
  estimatedCostNgn: number;
}