# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TARA Travel Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  seatPreference?: string | null;
  /** @nullable */
  mealPreference?: string | null;
  /** @nullable */
  airlinePreference?: string | null;
  walletBalance: number;
  taraPoints: number;
  onboardingComplete: boolean;
  plan: string;
  createdAt: Date;
}